+++ tests/integration_tests.rs
@@ -48,6 +48,9 @@ script_test!(vm_interfaces, "vm_interfaces.bxs");
 script_test!(vm_on_missing_method, "vm_on_missing_method.bxs");
 script_test!(vm_imports_alias, "vm_imports_alias.bxs");
 script_test!(vm_imports_no_alias, "vm_imports_no_alias.bxs");
+script_test!(vm_operators, "vm_operators.bxs");
+script_test!(vm_continue, "vm_continue.bxs");
+script_test!(vm_array_sparse, "vm_array_sparse.bxs");
 
 #[test]
 #[should_panic(expected = "must implement abstract method f")]